+++ src/DataFactory/CompositeDataFactory.js
@@ -9,10 +9,10 @@ const RecordDataFactory = require('./RecordDataFactory')
 const VariantDataFactory = require('./VariantDataFactory')
 const EventDataFactory = require('./EventDataFactory')
 const Bls12381DataFactory = require('./Bls12381DataFactory')
+const CallDataFactory = require('./CallDataFactory')
 
 class CompositeDataFactory {
     constructor() {
-        const variantFactory = new VariantDataFactory(this)
         this._factories = [
             new PrimitiveDataFactory(this),
             new ListDataFactory(this),
@@ -20,9 +20,10 @@ class CompositeDataFactory {
             new MapDataFactory(this),
             new TupleDataFactory(this),
             new RecordDataFactory(this),
-            new EventDataFactory(this, variantFactory),
-            variantFactory,
+            new EventDataFactory(this),
+            new VariantDataFactory(this),
             new Bls12381DataFactory(this),
+            new CallDataFactory(this),
         ]
     }
 
@@ -37,7 +38,7 @@ class CompositeDataFactory {
 
     create(type, value) {
         const factory = this._factories.find(f => f.supports(type))
-        assert(factory, `Unsupported type: ${JSON.stringify(type)}`)
+        assert(factory, `Unsupported type "${type.name}"`)
 
         return factory.create(type, value)
     }