+++ src/TypeResolver.js
@@ -82,14 +82,14 @@ class TypeResolver {
         return this.resolveType(funcAci.returns)
     }
 
-    getEventType(contract) {
+    getEventType(contract, topics) {
         const aci = this.getNamespaceAci(contract)
 
         if (!aci.hasOwnProperty('event')) {
             throw new TypeResolveError('Missing event declaration')
         }
 
-        return FateTypeEvent(this.resolveType(aci.event))
+        return FateTypeEvent(this.resolveType(aci.event), topics)
     }
 
     isCustomType(type) {
@@ -188,7 +188,7 @@ class TypeResolver {
             return FateTypeAccountAddress()
         }
 
-        if (key === 'contract_address') {
+        if (key === 'contract_pubkey') {
             return FateTypeContractAddress()
         }
 
@@ -301,7 +301,7 @@ class TypeResolver {
         }
 
         if (namespaceData.name === type) {
-            return ['contract_address', []]
+            return ['contract_pubkey', []]
         }
 
         const def = [