+++ src/types/FateOracleAddress.js
@@ -2,7 +2,7 @@ const FateAddress = require('./FateAddress')
 
 class FateOracleAddress extends FateAddress {
     constructor(value) {
-        super(value, 'oracle_address', 'ok')
+        super(value, 'oracle_pubkey', 'ok')
     }
 }